refactor: move tx input data into their own components
