+++ packages/app-explorer/src/systems/Transaction/component/TxInput/types.ts
@@ -1,0 +1,6 @@
+import type { GQLTransactionItemFragment } from '@fuel-explorer/graphql/sdk';
+import type { CardProps } from '@fuels/ui';
+
+export type TxInputProps = CardProps & {
+  input: NonNullable<GQLTransactionItemFragment['inputs']>[number] | undefined;
+};
